feat: add automatic shutdown of a stream after a specified delay
We need to progam our own logic to determine if an HLS stream is not viewed anymore because the library does not emit any event (`preConnect`, `postConnect`, `doneConnect`, etc.) for the HLS streams as it does for RTMP or FLV streams.

Our simple way to do it is to assume the HLS stream is not used anymore if we didn't received an HTTP request for any file of the HLS stream in 5 minutes.
We then make a request to the internal API to close the stream. This will stop the HLS transmuxing process and close the socket receiving the video stream from the camera.